+++ src/matterbridgeDynamicPlatform.ts
@@ -85,4 +85,19 @@ export class MatterbridgeDynamicPlatform {
   async registerDevice(device: MatterbridgeDevice) {
     await this.matterbridge.addBridgedDevice(this.name, device);
   }
+
+  /**
+   * Unegisters a device registered with the Matterbridge platform.
+   * @param {MatterbridgeDevice} device - The device to unregister.
+   */
+  async unregisterDevice(device: MatterbridgeDevice) {
+    await this.matterbridge.removeBridgedDevice(this.name, device);
+  }
+
+  /**
+   * Unegisters all devices registered with the Matterbridge platform.
+   */
+  async unregisterAllDevices() {
+    await this.matterbridge.removeAllBridgedDevices(this.name);
+  }
 }